File Restructure and Login/Register routes
In Create User, there is a placeholder true which is for the displaywishlist input, I will change this later but for now all users being registered will have their displaywishlist is true.

Register User Post route is not working as of now, there is no input validation for it so any incorrect inputs will crash the server. Please use the already existing users from npm run seed for any testing